// src/data/terrain-types.ts
export enum TerrainType {
  PLAINS = 'PLAINS',
  HILLS = 'HILLS',
  MOUNTAINS = 'MOUNTAINS',
  FOREST = 'FOREST',
  DEEP_FOREST = 'DEEP_FOREST',
  WETLAND = 'WETLAND',
  DESERT = 'DESERT',
  TUNDRA = 'TUNDRA',
  COAST = 'COAST',
  RIVER_VALLEY = 'RIVER_VALLEY'
}

export enum BiomeType {
  TEMPERATE = 'TEMPERATE',
  BOREAL = 'BOREAL',
  MEDITERRANEAN = 'MEDITERRANEAN',
  STEPPE = 'STEPPE',
  ARCTIC = 'ARCTIC',
  ARID = 'ARID',
  TROPICAL = 'TROPICAL'
}